REFAC: Use sha instead of stat.st_mtime to check for changed files
Sometimes, an output file cannot be loaded because the file is not
yet entirely written (LoadObjectError)
but the file is considered as checked and it is never loaded again.
The LazyAnalysisModel used to check the modification time of files
to detect changes, maybe it is not precise enough.
So we try using hashlib.sha1 instead.