ENH: Load the color map of labelled sulci at anatomist init + New class ColorMap with constants
The color map of sulci graph is specific in Anatomist : a nomenclature
file stored in brainvisa-share directory must be loaded before the graph
or a method updateColors() have to be called on the graph if it was already loaded.
This specificity is hidden in pyanatomist backend to keep the api simple.