PERF: For AC/PC points, use aims to apply a translation to the cross meshes
instead of creating a referential and loading the transformation in anatomist.
Indeed, the referentials are not deleted automatically when they are not used anymore.
So all referentials created for the AC/PC points of each subject were never deleted,
which really slowed the application when there was a lot of subjects.